fix(fd2): updates crop list after each transplanting (#230)
**Pull Request Description**

Causes the Crop list in the Transplanting form to update after each
transplant operation. This is necessary because if all of the trays of
the crop are transplanted, then the crop should no longer be listed.

Closes #225 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.