fix(fd2): archives tray seeded plant assets when fully transplanted (#231)
**Pull Request Description**

When a tray seeded plant asset has been fully transplanted its `TRAYS`
inventory will be 0. The changes in this PR ensure that the plant asset
is archived when this happens.

Closes #220 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.