fix(fd2): removes unnecessary quantity deletes (#232)
**Pull Request Description**

When a log referencing quantities is deleted those quantities are also
deleted. Thus, if a transaction fails and deletes a log with quantities,
it is not necessary to delete the quantities. This PR updates:

- the `runTransaction` function so that `undo` methods have the
information they need to skip these deletes.
- the Direct Seeding and Transplanting forms to use the information to
avoid errors caused by attempting to delete quantities that were already
deleted when the log was deleted..

Closes #218 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.