fix(fd2): enforces asset and log naming conventions (#235)
**Pull Request Description**

The naming conventions for logs and assets as described in
`docs/DataModel.md` are now enforced in the code.

Closes #208 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.