feat(dev): enables launching dev environment via GitPod (#236)
**Pull Request Description**

This PR makes it possible to launch the FarmData2 development
environment via GitPod. When launched this way there are no dependencies
as everything runs within the cloud and can be accessed via the browser.

Closes #77 
Partially addresses #4 by completing the INSTALL documentation.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.