feat(docs): updates the core documentation (#239)
**Pull Request Description**

This PR updates the core project documentation that would be relevant to
visitors to the repo (e.g. README, CODE_OF_CONDUCT, CONTRIBUTING, etc).
It also creates an documentation index that is the starting point for
all of the FarmData2 documentation.

Very rough notes for documentation of guides for working on components,
libraries, entry points and infrastructure are also added. These need
substantial additional work and separate issues will be created.

- Closes #4 
- Closes #5 
- Closes #6 
- Closes #7
- Closes #8 
- Closes #237
- Closes #240

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.