fix: builds single css file for each module (#246)
**Pull Request Description**

This PR changes the vite builds so that there is a single `css` file for
each module. Vite was splitting some of the CSS into component specific
files (e.g. `CropSelector/CropSelector.css`) Then to have this styling
work within an entry point that `css` file would need to be added to the
`*.libraries.yml` file so that Drupal makes it available to the page.
This was complicated, and a bit unpredictable because it was unclear why
or when vite would do this. So, instead a single `css` file is now
created in `style/style.css` and that file is included in the
`*.libraries.yml` for each entry point. That introduces a small
inefficiency as all of the CSS is loaded on each entry point even if it
is not used. But this should be a small overhead.

Closes #202 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.