fix(docs): corrected typos in codebase.md (#252)
**Pull Request Description**
Made the following changes:

From: "Dependencies ... are manages with npm" 
To: "Dependencies ... are managed with npm"

From: "...allowing the scripts can be run from any working directory" 
To: "...allowing the scripts to be run from any working directory"

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Signed-off-by: sidlamsal <98344853+sidlamsal@users.noreply.github.com>