fix(docs): adds a missing coma to line 27 of README.md (#253)
Added a comma to correct grammar.

**Pull Request Description**

Adds a comma to line 27 of README.md

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Signed-off-by: Shahir Ahmed <ahmeds@dickinson.edu>