fix(docs) updates port number for FarmData2 environment access (#255)
**Pull Request Description**

Updated the documentation to correct the port number for accessing the
FarmData2 development environment from 6109 to 6901, ensuring accurate
setup instructions. Verified the new port ensures proper access.

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Signed-off-by: Shahir Ahmed <ahmeds@dickinson.edu>