fix(docs): adds additional documentation for creating an SSH key (#258)
**Pull Request Description**

Added a link to GitPod's guide to create a SSH key in the event that the
user does not have one.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Signed-off-by: sidlamsal <98344853+sidlamsal@users.noreply.github.com>