fix(comp): adds prop to SoilDisturbance component to hide area field
**Pull Request Description**

Added `includeArea` prop to the `SoilDisturbance` component to allow the area field to be hidden.

Also added an entry point for an example page for the `SoilDisturbance` component.

closes #238 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.