fix(dev): uses vite to build e2e tests (#268)
**Pull Request Description**

Modifies the cypress and vite configurations so that e2e tests are built
using Vite instead of webpack. This allows the e2e test files to use the
aliases (e.g. `@libs/farmosUtil/farmosUtil`) to import libraries.

Closes #267

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.