fix(test): validates form data in submission e2e test (#270)
**Pull Request Description**

Uses a `spy` to validate the data that is passed to `lib.submitForm`
when the Submit button is clicked. This includes a number of other
changes that were needed to make that `spy` work. The details are
documented in #170.

Partially addresses #170 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.