fix(test): improves direct_seeding.submission.e2e.cy.js tests (#269)
**Pull Request Description**

Adds a test using a spy on the submitForm function to check that it was
called with the correct data.


Partially addresses #170 

Co-authored-by: Shahir Ahmed
<98346408+Shahir-47@users.noreply.github.com>

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.