fix(test): Direct seeding submission test (#273)
**Pull Request Description**

Adds a test using a spy on the submitForm function to check that it was
called with the correct data.

Partially addresses https://github.com/FarmData2/FarmData2/issues/170

This PR essentially un-reverts #269 which was reverted by #272 so that
the Co-author footers can be moved to the correct location.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---

Co-authored-by: Shahir Ahmed
<98346408+Shahir-47@users.noreply.github.com>
Co-authored-by: sidlamsal <98344853+sidlamsal@users.noreply.github.com>

---------

Co-authored-by: sidlamsal <lamsals@dickinson.edu>