fix(comp): soil disturbance fix (#280)
**Pull Request Description**

Changes the `isValid()` computed property of the SoilDisturbance
component to correctly output the validity based on if area is or is not
included. Also, adds a test to the `SoilDisturbance.events.comp.cy.js`
file to check that `isValid()` works accordingly.

Closes #276 
---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Co-authored-by: Shahir Ahmed <98346408+Shahir-47@users.noreply.github.com>
Co-authored-by: sidlamsal <98344853+sidlamsal@users.noreply.github.com>