feat(dev): improves entry point creation and maintainence (#251)
**Pull Request Description**

This pull requests adds the documentation for the creation and
maintenance of entry points. In addition, it updates the
`addEntrypoint.bash` script and the associated templates.

Closes #243
Closes #189
Addresses part of #170

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.