refactor(fd2): refactors tray seeding App.vue (#281)
**Pull Request Description**

Refactors the Tray Seeding endpoint to follow the template pattern in
`/bin/templates/endpoint`. Also:
  - enhances some of the Tray Seeding e2e tests.
  - formats some files to meet prettier style guide

Addresses part of #277.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.