refactor(fd2): modifies direct seeding App.vue to match new templates (#285)
**Pull Request Description**

Refactors the Direct Seeding endpoint to follow the template pattern in
`/bin/templates/endpoint`. Also enhances some of the Direct Seeding e2e
tests.

Addresses part of #27

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.