feat(dev): improves addEntrypoint.bash script (#289)
**Pull Request Description**

Improves the `addEntrypoint.bash` script by:
- Improving command line flag processing.
- Adding `--help` and usage instructions
- Adding `--min` flag to create a minimal entry point.
 
Closes #282 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.