fix(examples): moves example entry point to examples main menu (#293)
**Pull Request Description**

The Example Entry Point example was appearing on the FD2 Examples ->
Components menu. This PR moves it to the FD2 Examples menu instead.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.