fix(examples): update index of examples (#294)
**Pull Request Description**

Adds `PicklistBase`, `SelectorBase `, and `SoilDisturbance `, to the
list of examples at http://farmos/fd2_examples/component_examples.

closes #259 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.