fd2(fix): removes area from transplant form (#271)
**Pull Request Description**

Modifies the Transplanting form to not include area and adds unit and e2e tests.

Closes #265 #223 #222 #170

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---

Co-authored-by: sidlamsal <98344853+sidlamsal@users.noreply.github.com>