fix(dev): updates documentation for submit tests in entry point templates (#291)
**Pull Request Description**

Updates the documentation in the following entry point templates to
better describe the goals of the tests and how they check the
correctness of the code:
- `lib.submit.unit.cy.js` 
- `entry_point.submission.e2e.cy.js`

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.