fd2(fix): adds bed feet to soil disturbance log in direct seeding (#295)
**Pull Request Description**

Adds the `bedFeetQuanitity` to the soil disturbance activity log for
direct seeding and updates the `lib.submit.unit.cy.js` test to check
that `bedFeetQuanitity` is referenced in the soil disturbance log.

closes #286 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.