adds soil disturbed for transplanting and modifies test  (#296)
**Pull Request Description**

Adds the `transplantingBedFeetQuanitity` to the soil disturbance
activity log for transplanting and updates the `lib.submit.unit.cy.js`
test to check that `transplantingBedFeetQuanitity ` is referenced in the
soil disturbance log.

closes #287  

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.