refactor(fd2): updates transplanting to match new templates (#299)
**Pull Request Description**

Updates the transplanting entry point and tests to match the new entry
point templates. This simplifies the code and fixes an issue with
enabling / disabling the submit and reset button. Also touches up a few
things in the tray seeding and direct seeding entry points that were
missed when refactoring them.

Closes #277 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.