fd2(fix): changes the styling of the submit/reset buttons (#302)
**Pull Request Description**

Styles the buttons so that the reset button is a fixed size and goes all
the way to the right of the page. The Submit button should then expand
to fill the remainder of the space.

Closes #219 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.