fix(fd2): clears cashed values when terms/assets added to selectors (#300)
**Pull Request Description**

Changes `SelectorBase` component to emit event on `+` instead of changing URL.  Then modifies all parent components to handle that event.  This allows them to clear the relevant cached values and then redirect to the desired URL.

The following components were updated:

- `components/CropSelector`
- `components/LocationSelector`
- `components/TraySizeSelector`
- `components/EquipmentSelector`

Additionally, adds or modifies the component tests to confirm that
clicking the `+` clears the cached items.

Closes #103

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.