fix(lib): adds parent 'Category' asset for all equipment categories (#303)
**Pull Request Description**

This makes the equipment categories all have a common parent 'Category'.
This makes it possible to include equipment without a category because
we can omit any assets with this parent rather than omitting any assets
that do not have a parent.

Closes #298 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.