fix(test): fixes the equipment selector test after new db release (#304)
**Pull Request Description**

Fixes the equipment selector content component test after the new
v3.3.5. sample database release

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.