fix(lib): handles null for plantAsset parameter in farmosUtil.js: createSoilDisturbance (#306)
**Pull Request Description**

This PR handles null for `plantAsset` parameter in `farmosUtil.js:
createSoilDisturbance`. In addition, a test is added for creating a soil
disturbance without a `plantAsset`.

Closes #233 
---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.