fix(fd2): clarifies invalid feedback text on tray seeding form (#307)
**Pull Request Description**

The invalid feedback text is modified to be

- "Trays must be a positive number."
- "Seeds/Cell must be a positive number."

Closes #102
---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.