docs: makes components and libraries docs use markdown table (#309)
**Pull Request Description**

Changes the components and libraries page to use markdown table which
has two columns for "Name" and "Description". The content would be the
same as it currently is, just laid out in a table instead of a bulleted
list.

Closes #39

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.