fix(dev): adds docs build step to setup.bash (#310)
**Pull Request Description**

Add to the `setup.bash` script a step that builds the documentation of
the Components and Libraries.

Closes #261 
---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.