docs: fixes capitalization of addEntrypoint.bash comamnd (#313)
**Pull Request Description**

The P should be lower case in the `addEntrypoint.bash` command

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Signed-off-by: Shahir Ahmed <ahmeds@dickinson.edu>