fix(fd2): adds popup for + button in SelectorBase component (#305)
**Pull Request Description**

Changes the `+` button in `SelectorBase` to display a popup `iframe` containing the farmOS form for adding the appropriate type of item (field, bed, structure, crop, etc).

closes #152 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.