feat(fd2): adds the `WinterKill` component (#314)
**Pull Request Description**

Creates a component for indicating if a Cover Crop Seeding will be
winter killed. Additionally, it creates an example page in
`farm_fd2_examples` for the component and a set of component tests for
the component.

Closes #311 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------