test: patches tests for sampleDB update to 3.4.0 (#330)
**Pull Request Description**

Patches the tests to pass with the sampleDB update that adds direct
seedings, transplantings and some tillage equipment.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.