fix(comp): restyles the MultiCropSelector component and fixes a small bug (#332)
**Pull Request Description**

This pull request changes it to the style similar to the way the
equipment selector is displayed:

Also fixes a fug that occurred when the prop was used to remove
the crops when the input was required.

Closes #323 
Closes #329 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Co-authored-by: Shahir-47 <ahmeds@dickinson.edu>