fix(fd2): patches the equipment selector (#333)
**Pull Request Description**

Patches the EquipmentSelector and adds a test for the case when the
component is required and the last piece of equipment is removed via a
change to the prop. The component now emits `valid` with a `false`
payload in this circumstance.

Closes #328 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.