fix(fd2): changes required and delete behavior in multiple selectors  (#338)
**Pull Request Description**

Updates `MultiCropSelector` and `EquipmentSelector` so that when
multiple items are selected, only the top item has `*` and all items
have the delete button. With only one selected item, the top item has
`*` but no delete button. Additional testing for the 2 components have been added.

closes #320 #319 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.