feat(fd2): adds entry point for soil disturbance events (#318)
**Pull Request Description**

Adds:
- soil disturbance entry point
- lib.js including tests
- e2e tests

Closes #249 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Co-authored-by: Shahir Ahmed
[98346408+Shahir-47@users.noreply.github.com](mailto:98346408+Shahir-47@users.noreply.github.com)