fix(dev): enhances the `runAlltests.bash` script (#350)
**Pull Request Description**

Causes the script to only print:
- Each "Running ... (x of y)" line as it goes for status updates.
- Any failed test output.
- The results table at the end of each collection of tests.

Closes #108 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.