fix(test): patches winterkill test for added log type (#351)
**Pull Request Description**

Patches this test to account for the addition of the
`seeding_cover_crop` log type being added to the log.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.