refactor(fd2): improves performance of fetching plant assets by location (#354)
**Pull Request Description**

Adds an API endpoint to farmOS for fetching plant assets by location and
refactors the `getPlantAssets` function in `farmosUtil.js` to use it.
This significantly improves the performance of that function and
improves the UX of the Soil Disturbance form because the "Termination
Event" check box appears/disappears more quickly.

Adds some draft documentation about creating Views.

Refactors some of the e2e tests to:
- Patch for a new location (H) added in the sample database (v3.4.1)
- Include some missing assertions for sticky/non-sticky content in cover
crop tests
- reorganize some assertions for sticky/non-sticky content in soil
disturbance tests
- Ensure that locations and beds that are intended to be empty remain
empty.

Closes #340 
Closes #195 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Signed-off-by: Grant Braught <braught@dickinson.edu>