fix(examples): fixes indentation error in .yml file (#362)
**Pull Request Description**

Fixes an indentation error in the examples `routing` yml file that was
introduced when resolving merge conflicts.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.