feat(examples): adds NumericInput example (#352)
**Pull Request Description**

Adds example page for `NumericInput` component. 

Closes #355 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Signed-off-by: Grant Braught <braught@dickinson.edu>
Co-authored-by: Grant Braught <braught@dickinson.edu>