feat(components): adds SortOrderButton component and adds it to PickListBase (#361)
**Pull Request Description**

- Creates a new `SortOrderButton` component, which is a button that can
be used in the column headers of tables to control/indicate which column
the table is sorted by. The button will indicate with an icon either up
or down for sort in increasing or sort in decreasing.
-  Creates an example page for the `SortOrderButton` component.
- Modifies the `PickListBase` component to use a `SortOrderButton`
component as the column headers.
- Modifies the `PicklistBase` example page to test the newly added
`SortOrderButton` component.
- Writes meaningful component tests for `PicklistBase` component.

Closes #197
Closes #188 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.