fix(dev): added docker daemon check (#379)
**Pull Request Description**

Closes #378 by running a quick `docker info` to check if the Docker
daemon is running.

Closes #378

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Signed-off-by: Ty Chermsirivatana <chermsit@dickinson.edu>