feat(dev): adds fetchPR.bash and pushPR.bash scripts (#381)
**Pull Request Description**

Adds scripts for fetching a PR branch and pushing a PR branch. This
allows a maintainer who has been given permission to add commits to a PR
branch to easily fetch the branch, make changes, commit, and push the PR
branch.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.