Refactor PicklistBase to Use Map for Selected Rows and Fix Sorting Issues (#385)
**Pull Request Description**

This PR refactors the `PicklistBase.vue` component to store selected
rows in a `Map`, resolving the sorting-related issues described in #383.
The internal `pickedRows` array is kept for `BFormSelect` compatibility,
while the external `picked` prop now uses a `Map` for cleaner selection
handling. Updates include:
- `PicklistBase.vue` refactor
- Changes to `TransplantingPicklist.vue`
- Updates to example pages and associated tests
- Modifications to end-to-end transplanting tests

Closes #383 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.