rightside of SelectorBase and NumericInput button corners are rounded (#392)
**Pull Request Description**

This was due to a `class="has-validation"` class assignment missing from
the `inputGroup` element

Closes #391 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Co-authored-by: RolandLocke <roland.w.locke@gmail.com>