fix(comp): ActivePlantAssetPicklist - shows BedPicker and Picklist as appropriate (#501)
**Pull Request Description**

Changed it to it.only in active plant asset picklist to create PR. Will
remove after.

Closes #490

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.