fix(comp): PicklistBase - visual styling bug (#511)
**Pull Request Description**

Fixed visual styling bug in PicklistBase component that would occur when
showValidityStyling prop was switched to true. Green border would show
when it should've been the normal border color.

Closes #491 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.