fix(comp): ActivePlantAssetPicklist - patch validity computation (#504)
**Pull Request Description**

Pull request for ActivePlantAssetPicklist Validity

Closes #489 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Co-authored-by: braughtg <braught@dickinson.edu>