Required row update (#515)
**Pull Request Description**

Updates eactive plant asset picklist example page so that required is
forced to be true when requiredRow is true.


Closes #489 
---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Co-authored-by: braughtg <braught@dickinson.edu>