added prop for includeemptybeds (#514)
**Pull Request Description**

Added includeEmptyBeds prop to ActivePlantAssetPicklist to control bed
filtering
When includeEmptyBeds is false, only show beds with active plant assets.
When true, show all beds in the location. Uses PickerBase directly for
cleaner implementation.

Closes #513 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.