fix(comp): ActivePlantAssetPicklist - select all beds if no active plant assets (#518)
**Pull Request Description**

Select all beds if beds exist and there are no active plant assets in
the location.

Closes #499 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.