fix(comp): ActivePlantAssetPicklist - update affected area computation (#516)
**Pull Request Description**

Updated the way the affected area estimates are being computed using a
computed property and a watcher.

Closes #500 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

---------

Co-authored-by: braughtg <braught@dickinson.edu>