fix(fd2): patches soil disturbance submission test for empty beds (#522)
**Pull Request Description**

Adds an empty bed selection to the submission test for the case with
active plant assets.

Closes #495 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.