dev: .devcontainer patch
Patches the .devcontainer to reveal the terminal in which VSCode tasks are running.  This appears to be a change in how GitHub codespaces execute the VSCode tasks.  If the terminal was not revealed the first task would fail and no further tasks would run.  Revealing the terminal window solved this issue.

Signed-off-by: Grant Braught <braught@dickinson.edu>